ODE solver from DifferentialEquations.jl (#12)
The ODE solver is from DifferentialEquations.jl and test against the original solver from dftatom.
All Adams-Bashforth Explicit Methods can be used. The VCABM5 algorithm is hard coded for the moment (TODO: benchmark on the efficiency and numerical stability). In terms of the stability (where integration start explode), the VCABM5 is a bit better than the implicit AB5 of dftatom.
The code base is refactoring by moving all dftatom fortran wrapper and dftatom mimic implementation to the legacy folder.